+++ lib/td_cache/permissions_config.ex
@@ -48,7 +48,8 @@ defmodule TdCache.PermissionsConfig do
     :manage_confidential_structures,
     :manage_ingest_relations,
     :view_data_structures_profile,
-    :view_quality_rule
+    :view_quality_rule,
+    :manage_quality_rule_implementations
   ]
 
   def permissions, do: @permissions